backend/requirements.txt
fastapi[standard]
sqlalchemy
psycopg2
confluent-kafka
requests
pytz
snowflake-connector-python
pandas
passlib
pyjwt
passlib[bcrypt]
python-jose